feat: 修改图表示例
