+++ src/config/routerData.ts
@@ -280,6 +280,26 @@ export default {
 					permission: ["C", "R", "U", "D"],
 					children: [],
 				},
+				{
+					id: 906,
+					path: "echart",
+					component: "demo/chart/Echart",
+					name: "message.menu.demoEchart",
+					title: "message.menu.demoEchart",
+					icon: "icon-appstore",
+					isLink: 0,
+					isIframe: 0,
+					address: "",
+					isAffix: 1,
+					isKeepAlive: 1,
+					isDisable: 0,
+					isHide: 0,
+					isHideSubMenu: 1,
+					isMobile: 0,
+					roles: ["admin", "system"],
+					permission: ["C", "R", "U", "D"],
+					children: [],
+				},
 			],
 		},
 		{
