+++ src/plugins/i18n/modules/zh-cn.ts
@@ -49,6 +49,7 @@ export default {
 		demoMapOne: "CV分割地图",
 		demoMapTwo: "CV拖拽",
 		demoUa: "UA&IP",
+		demoEchart: "图表Echart",
 		testLink: "链接",
 		testIframe: "Iframe",
 	},
